Move ARXReIdentificationRiskFactory to analyzer package as it seem to belong there
